requirements.txt
celery~=5.2.3
celery-redbeat~=2.0.0
fakeredis~=1.7.1
fastapi~=0.73.0
kafka-python~=2.0.2
requests~=2.27.1
psycopg2~=2.9.3
pydantic~=1.9.0
pytest~=7.0.0
typer
